pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "supre-robot-sdk"
version = "0.1.0"
description = "Standalone hardware SDK for Supre dual-arm robots"
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["pyyaml"]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

